+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/Account.kt
@@ -1,11 +1,12 @@
 package com.dllewellyn.coinbaseapi.models
 
 import com.dllewellyn.coinbaseapi.models.currency.SupportedCurrency
+import java.math.BigInteger
 
 data class Account(
     val currencyValue: SupportedCurrency,
-    val balance : Double,
-    val available : Double,
-    val hold : Double,
-    val uid : String
+    val balance: BigInteger,
+    val available: BigInteger,
+    val hold: BigInteger,
+    val uid: String
 )
