+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/ExchangeRates.kt
@@ -1,7 +1,8 @@
 package com.dllewellyn.coinbaseapi.models
 
 import com.dllewellyn.coinbaseapi.models.currency.CryptoCurrency
+import java.math.BigInteger
 
-data class ExchangeRates(val cryptoCurrency: CryptoCurrency, val map : Map<String, Double>) {
+data class ExchangeRates(val cryptoCurrency: CryptoCurrency, val map : Map<String, BigInteger>) {
     fun filterForCurrency(currency : String) = map[currency]
 }
