+++ rxadapter/src/main/kotlin/com/dllewellyn/coinbaseapi/websocket/internal/models/SnapshotResponseApi.kt
@@ -3,6 +3,7 @@ package com.dllewellyn.coinbaseapi.websocket.internal.models
 import com.dllewellyn.coinbaseapi.models.trade.CurrencyBuyAndSell
 import com.dllewellyn.coinbaseapi.models.EventResponse
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class SnapshotResponseApi(
     @SerializedName("asks") val asks: List<List<String>>,
@@ -16,8 +17,8 @@ data class SnapshotResponseApi(
             CurrencyBuyAndSell(
                 product_id.productIdToPair().first(),
                 product_id.productIdToPair().last(),
-                bids.first().first().toDouble(),
-                asks.first().first().toDouble()
+                BigInteger(bids.first().first()),
+                BigInteger(asks.first().first())
             )
         )
     }
