+++ rxadapter/src/main/kotlin/com/dllewellyn/coinbaseapi/websocket/internal/models/UpdateResponseApi.kt
@@ -5,6 +5,7 @@ import com.dllewellyn.coinbaseapi.models.currency.CurrencyValue
 import com.dllewellyn.coinbaseapi.models.EventResponse
 import com.google.gson.annotations.SerializedName
 import java.lang.IllegalArgumentException
+import java.math.BigInteger
 
 data class UpdateResponseApi(
     @SerializedName("changes") val changes: List<List<String>>,
@@ -22,12 +23,15 @@ data class UpdateResponseApi(
                 else -> throw IllegalArgumentException()
             }
 
-            returnList.add(EventResponse.Level2Update(
-                CurrencyValue(
-                    product_id.productIdToPair().first(),
-                    product_id.productIdToPair().last(),
-                    it[1].toDouble()
-                ), buyOrSell, it[2].toDouble()))
+            returnList.add(
+                EventResponse.Level2Update(
+                    CurrencyValue(
+                        product_id.productIdToPair().first(),
+                        product_id.productIdToPair().last(),
+                        BigInteger(it[1])
+                    ), buyOrSell, it[2].toDouble()
+                )
+            )
         }
 
         return returnList
